Modularise notion-sync into per-database modules
Split monolithic sync.ts, sqlite-writer.ts, schema.ts, and types.ts
into self-contained database modules under src/databases/:
- transactions/ (types, mapper, schema, writer)
- entities/ (types, mapper, schema, writer, lookup)
- inventory/ (types, mapper, schema, writer)
- budgets/ (types, mapper, schema, writer)
- wish-list/ (types, mapper, schema, writer)

Each module owns its row type, Notion-to-row mapper, CREATE TABLE
DDL, and SQLite upsert function. Barrel index.ts re-exports all.

Shared infrastructure extracted:
- cursor.ts: saveCursor/loadCursor for incremental sync state
- test-helpers.ts: richText() and makePage() used across test files

sync.ts is now a thin orchestrator that imports from per-database
modules. types.ts re-exports row types for backwards compatibility.

All 54 tests pass. No functional changes.
