+++ services/notion-sync/src/databases/budgets/index.ts
@@ -1,0 +1,4 @@
+export { type BudgetRow } from "./types.js";
+export { createBudgetsTable } from "./schema.js";
+export { upsertBudgets } from "./writer.js";
+export { mapBudget } from "./mapper.js";
