+++ services/notion-sync/src/databases/entities/index.ts
@@ -1,0 +1,5 @@
+export { type EntityRow } from "./types.js";
+export { createEntitiesTable } from "./schema.js";
+export { upsertEntities } from "./writer.js";
+export { mapEntity } from "./mapper.js";
+export { buildEntityLookup } from "./lookup.js";
