+++ services/notion-sync/src/databases/entities/types.ts
@@ -1,0 +1,12 @@
+/** Flattened entity row for SQLite (mirrors Notion Entities database). */
+export interface EntityRow {
+  notionId: string;
+  name: string;
+  type: string | null;
+  abn: string | null;
+  aliases: string | null;
+  defaultTransactionType: string | null;
+  defaultCategory: string | null;
+  notes: string | null;
+  lastEditedTime: string;
+}
