+++ services/notion-sync/src/databases/inventory/index.ts
@@ -1,0 +1,4 @@
+export { type InventoryRow } from "./types.js";
+export { createInventoryTable } from "./schema.js";
+export { upsertInventory } from "./writer.js";
+export { mapInventoryItem } from "./mapper.js";
