+++ services/notion-sync/src/databases/inventory/schema.ts
@@ -1,0 +1,28 @@
+import type Database from "better-sqlite3";
+
+/** Create the home_inventory table. Idempotent. */
+export function createInventoryTable(db: Database.Database): void {
+  db.exec(`
+    CREATE TABLE IF NOT EXISTS home_inventory (
+      notion_id              TEXT PRIMARY KEY,
+      item_name              TEXT NOT NULL,
+      brand                  TEXT,
+      model                  TEXT,
+      item_id                TEXT,
+      room                   TEXT,
+      location               TEXT,
+      type                   TEXT,
+      condition              TEXT,
+      in_use                 INTEGER NOT NULL DEFAULT 0,
+      deductible             INTEGER NOT NULL DEFAULT 0,
+      purchase_date          TEXT,
+      warranty_expires       TEXT,
+      replacement_value      REAL,
+      resale_value           REAL,
+      purchase_transaction_id TEXT,
+      purchased_from_id      TEXT,
+      purchased_from_name    TEXT,
+      last_edited_time       TEXT NOT NULL
+    );
+  `);
+}
