+++ services/notion-sync/src/databases/inventory/types.ts
@@ -1,0 +1,22 @@
+/** Flattened home inventory row for SQLite (mirrors Notion Home Inventory). */
+export interface InventoryRow {
+  notionId: string;
+  itemName: string;
+  brand: string | null;
+  model: string | null;
+  itemId: string | null;
+  room: string | null;
+  location: string | null;
+  type: string | null;
+  condition: string | null;
+  inUse: boolean;
+  deductible: boolean;
+  purchaseDate: string | null;
+  warrantyExpires: string | null;
+  replacementValue: number | null;
+  resaleValue: number | null;
+  purchaseTransactionId: string | null;
+  purchasedFromId: string | null;
+  purchasedFromName: string | null;
+  lastEditedTime: string;
+}
