+++ services/notion-sync/src/databases/transactions/index.ts
@@ -1,0 +1,4 @@
+export { type TransactionRow } from "./types.js";
+export { createTransactionsTable } from "./schema.js";
+export { upsertTransactions } from "./writer.js";
+export { mapTransaction } from "./mapper.js";
