+++ services/notion-sync/src/databases/wish-list/index.ts
@@ -1,0 +1,4 @@
+export { type WishListRow } from "./types.js";
+export { createWishListTable } from "./schema.js";
+export { upsertWishList } from "./writer.js";
+export { mapWishListItem } from "./mapper.js";
