+++ services/notion-sync/src/databases/wish-list/schema.ts
@@ -1,0 +1,17 @@
+import type Database from "better-sqlite3";
+
+/** Create the wish_list table. Idempotent. */
+export function createWishListTable(db: Database.Database): void {
+  db.exec(`
+    CREATE TABLE IF NOT EXISTS wish_list (
+      notion_id        TEXT PRIMARY KEY,
+      item             TEXT NOT NULL,
+      target_amount    REAL,
+      saved            REAL,
+      priority         TEXT,
+      url              TEXT,
+      notes            TEXT,
+      last_edited_time TEXT NOT NULL
+    );
+  `);
+}
